+++ packages/adapters/src/adapters/Cornerstone3D/index.ts
@@ -13,12 +13,15 @@ import Length from "./Length";
 import PlanarFreehandROI from "./PlanarFreehandROI";
 import Probe from "./Probe";
 import UltrasoundDirectional from "./UltrasoundDirectional";
+import BaseAdapter3D from "./BaseAdapter3D";
 
 import * as Segmentation from "./Segmentation";
 import * as ParametricMap from "./ParametricMap";
 import * as RTSS from "./RTStruct";
+import KeyImage from "./KeyImage";
 
 const Cornerstone3DSR = {
+    BaseAdapter3D,
     Bidirectional,
     CobbAngle,
     Angle,
@@ -30,6 +33,7 @@ const Cornerstone3DSR = {
     Probe,
     PlanarFreehandROI,
     UltrasoundDirectional,
+    KeyImage,
     MeasurementReport,
     CodeScheme,
     CORNERSTONE_3D_TAG