+++ packages/dicomImageLoader/src/imageLoader/setPixelDataType.ts
@@ -1,0 +1,25 @@
+import getPixelDataTypeFromMinMax from '../shared/getPixelDataTypeFromMinMax';
+
+/**
+ * Helper function to set the right typed array.
+ * This is needed because web workers can transfer array buffers but not typed arrays
+ *
+ * Here we are setting the pixel data to the right typed array based on the final
+ * min and max values
+ */
+function setPixelDataType(imageFrame) {
+  const minValue = imageFrame.smallestPixelValue;
+  const maxValue = imageFrame.largestPixelValue;
+
+  const TypedArray = getPixelDataTypeFromMinMax(minValue, maxValue);
+
+  if (TypedArray) {
+    // @ts-ignore
+    const typedArray = new TypedArray(imageFrame.pixelData);
+    imageFrame.pixelData = typedArray;
+  } else {
+    throw new Error('Could not apply a typed array to the pixel data');
+  }
+}
+
+export default setPixelDataType;