+++ packages/polymorphic-segmentation/src/types/index.ts
@@ -1,0 +1,3 @@
+import type { PolySegConversionOptions } from './PolySegConversionOptions';
+
+export type { PolySegConversionOptions };